move K210Layer class to new file
